requirements.txt
setuptools~=53.0.0
PyYAML~=5.4.1